Don't test 128bit atomics on Windows with LLVM 3.9
LLVM 3.9 reworked the atomics lowering pass and as an effect it issues
a libcall to `__atomic_load` that we can't resolve right now. On Linux
we resolve that libcall to `libatomic`, but that is not available on
Mingw64 4.9. Later version do have it, but they currently run into
trouble with compiling LLVM on Win32.